# swagger.yaml
openapi: 3.0.0
info:
  title: TV Project API
  description: API документація для TV проекту
  version: 1.0.0
servers:
  - url: http://localhost:8080
  - url: https://tv-company-server.onrender.com
...
paths:
...
  /customers:
    get:
      security: 
        - bearerAuth: []
      tags: 
        - Customers
      operationId: getAllCustomers
      summary: Get list of all customers
      description: Endpoint to get a list of all customers
      responses:
        '200':
          description: Successfully retrieved list of customers
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Customer-res"
                  - type: object
                    properties:
                      message:
                        type: string
                        example: There are no agents in the database.
        "401":
          description: Not authorized
          
    post:
      security: 
        - bearerAuth: []
      tags: 
        - Customers
      operationId: createCustomer
      summary: Create a new customer
      description: Endpoint to create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer-req'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer-res"
        "401":
          description: Not authorized
        "409":
          description: Customer already exists
  /customers/{customerId}:
    get:
      security: 
        - bearerAuth: []
      tags: 
        - Customers
      operationId: getCustomerByID
      summary: Get a customer by ID
      description: Endpoint to get a customer by its ID.
      parameters:
        - in: path
          name: customerId
          required: true
          description: Customer ID
          schema:
            type: string
          example: "668c083ec1c3e8df54c6f669"
      responses:
        '200':
          description: Successfully retrieved the customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer-res'
        "401":
          description: Not authorized
        '404':
          description: Customer not found

    put:
      security: 
        - bearerAuth: []
      tags: 
        - Customers
      operationId: updateCustomerData
      summary: Update customer data
      description: Endpoint to update the data of an existing customer.
      parameters:
        - in: path
          name: customerId
          required: true
          description: Customer ID
          schema:
            type: string
          example: "668c083ec1c3e8df54c6f669"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer-req'
      responses:
        '200':
          description: Customer data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer-res'
        "401":
          description: Not authorized
        '404':
          description: Customer not found

    delete:
      security: 
        - bearerAuth: []
      tags: 
        - Customers
      operationId: deleteCustomer
      summary: Delete a customer
      description: Endpoint to delete a customer by its ID
      parameters:
        - in: path
          name: customerId
          required: true
          description: Customer ID
          schema:
            type: string
          example: "668c083ec1c3e8df54c6f669"
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found  
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer-req:
      type: object
      required: 
        - name
        - phone
        - contactPerson
        - bankDetails
      properties:
        name:
          type: string
          example: 'Customer fullname'
        phone:
          type: string
          example: "+380671234567"
        contactPerson:
          type: string
          example: 'Contact person full name'
        bankDetails:
          type: object
          required: 
            - bankName
            - customerCode
            - iban
          properties:
            bankName: 
              type: string
              example: "Bank name"
            customerCode: 
              type: string
              example: "12345678"
            iban:
              type: string
              example: "UA12345678901234567890"
    Customer-res:
      allOf:
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              example: "668c083ec1c3e8df54c6f669"
        - $ref: "#/components/schemas/Customer-req"    
...